Update ibcli to allow you to enable a view
Previous code only allowed you to add/modify a view disabled, however now you can add/modify a view to be enabled (or disabled=false, if you look at the code).  This means very little for adding a view, but you can now modify a view to be enabled or disabled from ibcli.